requirements.txt
flask
transformers
torch
pillow
python-docx
PyPDF2
hf_xet
